feat: chart 동적 기능 구현 완료
